rework priv bits into M_PRIV and item->data

+++ src/libtrx/include/libtrx/game/objects/traps/movable_block.h
@@ -10,5 +10,5 @@ void MovableBlock_HandleFlipMap(ROOM_FLIP_STATUS flip_status);
 void MovableBlock_SetPushPull(ITEM *item, bool enable);
 bool MovableBlock_IsPushPull(const ITEM *item);
 void MovableBlock_SetGravityFrames(ITEM *item, uint8_t frames);
-uint8_t MovableBlock_GetGravityFrames(const ITEM *item);
+uint16_t MovableBlock_GetGravityFrames(const ITEM *item);
 void MovableBlock_ActivateStack(const ITEM *base_item, XYZ_32 sector_pos);
